refactor: promote the TX press-gesture recognizer to wiring/
Mechanical move only, zero content change: git mv
panels/tx-ptt-gesture.ts -> wiring/tx-ptt-gesture.ts and its test, plus
the one import-line update in TxPanel.svelte. Content of both moved
files is byte-identical to what merged in PR #2125/#2128 (verified by
SHA-256).

Prepares MOR-1012: MobileRadioLayout will consume the recognizer, and
wiring/ is its natural home (precedent: double-click.ts).
