+++ frontend/src/components-v2/wiring/tx-ptt-gesture.ts
